+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 
 - Drastically increased spawn chances for *undefeated* trainers and lowered the bonus spawn chance of key trainers (e.g. leaders) gained by players reaching their level cap
 
+***Fixed***
+
+- Advancement defeat count trigger counting all instead of distinct defeats per trainer type
+
 ## [0.11.1-alpha] - 2024-08-31
 
 ***Fixed***